Rename utility files to use camel casing
