+++ src/util/filterTreeNodes.js
