+++ test/unit/specs/util/filterTreeNodes.spec.js
