+++ test/unit/specs/util/sortArray.spec.js
